backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
aiofiles==23.2.1
moviepy==1.0.3
langchain>=0.1.0
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
faiss-cpu>=1.7.0
langchain-community>=0.0.35
numpy<2.0
scipy>=1.11.3
pillow>=10.0.0
transformers>=4.35.0
torch>=2.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
# Tesseract OCR 설치 필요:
# - Windows: https://github.com/UB-Mannheim/tesseract/wiki 에서 설치
# - macOS: brew install tesseract tesseract-lang
# - Linux: sudo apt-get install tesseract-ocr libtesseract-dev tesseract-ocr-kor
# Google Gemini API 사용을 위해 필요
google-generativeai>=0.3.0